front end/requirements.txt
# requirements.txt

absl-py==2.2.2
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cmake==4.0.0
colorama==0.4.6
contourpy==1.3.2
cvzone==1.6.1
cycler==0.12.1
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
gast==0.4.0
geographiclib==2.0
geopy==2.4.1
google-auth==2.39.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
idna==3.10
imageio==2.37.0
imbalanced-learn==0.13.0
imblearn==0.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
joblib==1.4.2
keras==2.12.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
lime==0.2.0.1
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
narwhals==1.35.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.5
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==6.0.1
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
selenium==4.31.0
sentencepiece==0.2.0
shap==0.47.2
six==1.17.0
sklearn-compat==0.1.3
slicer==0.0.8
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.7
sympy==1.13.1
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.14.1
wsproto==1.2.0
xgboost==3.0.0
